refactor: 레거시 Profile을 디자인 시스템 Profile로 전환
